Modifies deployment script to unconditionally delete and re-upload keys.
The reason I think something like this would be better than how it
currently works is that it's very easy to end up with stale keys if
you delete an entire environment and recreate from scratch. However,
it's not nice that it deletes existing keys. Maybe the key names
should be appended with a (shortened) uuid so that they don't collide?

[This commit shouldn't be merged. Needs discussion with mega team]